Cleaned up integration tests a bit, added a new one
