+++ tests/integration/src/test/scala/dcos/metronome/integrationtest/utils/MetronomeFramework.scala
@@ -51,6 +51,7 @@ object MetronomeFramework {
     val uuid = UUID.randomUUID.toString
 
     lazy val httpPort = PortAllocator.ephemeralPort()
+    lazy val httpsPort = PortAllocator.ephemeralPort()
 
     // lower the memory pressure by limiting threads.
     val akkaJvmArgs = Seq(
@@ -82,6 +83,7 @@ object MetronomeFramework {
       "mesos_authentication_principal" -> "principal",
       "mesos_role" -> mesosRole,
       "http_port" -> httpPort.toString,
+      "https_port" -> httpsPort.toString,
       "zk" -> zkUrl,
       "zk_timeout" -> 20.seconds.toMillis.toString,
       "zk_connection_timeout" -> 20.seconds.toMillis.toString,
@@ -115,7 +117,7 @@ object MetronomeFramework {
 
       val cmd = Seq(java, "-Xmx1024m", "-Xms256m", "-XX:+UseConcMarkSweepGC", "-XX:ConcGCThreads=2") ++
         runtimeArguments ++ akkaJvmArgs ++
-        Seq(s"-DmarathonUUID=$uuid -DtestSuite=$suiteName", s"-Dmetronome.zk.url=$zkUrl", s"-Dmetronome.mesos.master.url=$masterUrl", s"-Dmetronome.framework.name=metronome-$uuid", s"-Dplay.server.http.port=$httpPort", "-classpath", cp, "-client", mainClass) // ++ args
+        Seq(s"-DmarathonUUID=$uuid -DtestSuite=$suiteName", s"-Dmetronome.zk.url=$zkUrl", s"-Dmetronome.mesos.master.url=$masterUrl", s"-Dmetronome.framework.name=metronome-$uuid", s"-Dplay.server.http.port=$httpPort", s"-Dplay.server.https.port=$httpsPort", "-classpath", cp, "-client", mainClass) // ++ args
 
       logger.info(s"Starting process in ${workDir}, Cmd is ${cmd}")
